todo/serializers.py
from rest_framework import serializers

from .models import Project, Todo


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = "__all__"


class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        exclude = ('is_active',)
